creating file to export each configuration of app
